requirements.txt
cssselect
Pillow
requests
lxml
